refactor: create utils module for shared helpers
