+++ src/prospector_holds/models/leader.py
@@ -11,6 +11,7 @@ see the MARC 21 Format for Bibliographic Data [1].
 """
 
 from ..settings import SCHEMA_JSON
+from ..utils import label_to_key
 
 
 """
@@ -20,20 +21,6 @@ LEADER_LITERALS = (
     'LEADER',
     'LDR',
 )
-
-
-def label_to_key(label):
-    """
-    Convert a label (of space-delimited words) into a dict key
-
-    Assume: label matches the following regex:
-        ^[a-zA-Z][-_ a-zA-Z0-9]*$
-    """
-    key = label
-    key = ' '.join(key.split('-'))
-    key = '_'.join(key.split(' '))
-    key = key.lower()
-    return key
 
 
 class Leader:
